# Implement GCM Seal and Open in Go