--- a/cipher.go
+++ b/cipher.go
@@ -457,11 +457,22 @@ func (g *cipherGCM) Seal(dst, nonce, plaintext, additionalData []byte) []byte {
 	// relying in the explicit nonce being securely set externally,
 	// and it also gives some interesting speed gains.
 	// Unfortunately we can't use it because Go expects AEAD.Seal to honor the provided nonce.
-	if C.go_openssl_EVP_CIPHER_CTX_seal_wrapper(ctx, base(out), base(nonce),
-		base(plaintext), C.int(len(plaintext)),
-		base(additionalData), C.int(len(additionalData))) != 1 {
-
-		panic(fail("EVP_CIPHER_CTX_seal"))
+	if C.go_openssl_EVP_EncryptInit_ex(ctx, nil, nil, nil, base(nonce)) != 1 {
+		panic(newOpenSSLError("EVP_EncryptInit_ex"))
+	}
+	var out_len, discard_len C.int
+	if C.go_openssl_EVP_EncryptUpdate(ctx, nil, &discard_len, baseNeverEmpty(additionalData), C.int(len(additionalData))) != 1 ||
+		C.go_openssl_EVP_EncryptUpdate(ctx, base(out), &out_len, baseNeverEmpty(plaintext), C.int(len(plaintext))) != 1 {
+		panic(newOpenSSLError("EVP_EncryptUpdate"))
+	}
+	if len(plaintext) != int(out_len) {
+		panic("cipher: incorrect length returned from GCM EncryptUpdate")
+	}
+	if C.go_openssl_EVP_EncryptFinal_ex(ctx, base(out[out_len:]), &discard_len) != 1 {
+		panic(newOpenSSLError("EVP_EncryptFinal_ex"))
+	}
+	if C.go_openssl_EVP_CIPHER_CTX_ctrl(ctx, C.GO_EVP_CTRL_GCM_GET_TAG, 16, unsafe.Pointer(base(out[out_len:]))) != 1 {
+		panic(newOpenSSLError("EVP_CIPHER_CTX_ctrl"))
 	}
 	runtime.KeepAlive(g)
 	return ret
@@ -469,7 +480,7 @@ func (g *cipherGCM) Seal(dst, nonce, plaintext, additionalData []byte) []byte {
 
 var errOpen = errors.New("cipher: message authentication failed")
 
-func (g *cipherGCM) Open(dst, nonce, ciphertext, additionalData []byte) ([]byte, error) {
+func (g *cipherGCM) Open(dst, nonce, ciphertext, additionalData []byte) (_ []byte, err error) {
 	if len(nonce) != gcmStandardNonceSize {
 		panic("cipher: incorrect nonce length given to GCM")
 	}
@@ -497,18 +508,33 @@ func (g *cipherGCM) Open(dst, nonce, ciphertext, additionalData []byte) ([]byte,
 		return nil, err
 	}
 	defer C.go_openssl_EVP_CIPHER_CTX_free(ctx)
-	ok := C.go_openssl_EVP_CIPHER_CTX_open_wrapper(
-		ctx, base(out), base(nonce),
-		base(ciphertext), C.int(len(ciphertext)),
-		base(additionalData), C.int(len(additionalData)), base(tag))
-	runtime.KeepAlive(g)
-	if ok == 0 {
-		// Zero output buffer on error.
-		for i := range out {
-			out[i] = 0
+
+	defer func() {
+		if err != nil {
+			// Zero output buffer on error.
+			for i := range out {
+				out[i] = 0
+			}
 		}
+	}()
+	if C.go_openssl_EVP_DecryptInit_ex(ctx, nil, nil, nil, base(nonce)) != 1 {
 		return nil, errOpen
 	}
+	if C.go_openssl_EVP_CIPHER_CTX_ctrl(ctx, C.GO_EVP_CTRL_GCM_SET_TAG, 16, unsafe.Pointer(base(tag))) != 1 {
+		return nil, errOpen
+	}
+	var out_len, discard_len C.int
+	if C.go_openssl_EVP_DecryptUpdate(ctx, nil, &discard_len, baseNeverEmpty(additionalData), C.int(len(additionalData))) != 1 ||
+		C.go_openssl_EVP_DecryptUpdate(ctx, base(out), &out_len, baseNeverEmpty(ciphertext), C.int(len(ciphertext))) != 1 {
+		return nil, errOpen
+	}
+	if len(ciphertext) != int(out_len) {
+		return nil, errOpen
+	}
+	if C.go_openssl_EVP_DecryptFinal_ex(ctx, base(out[out_len:]), &discard_len) != 1 {
+		return nil, errOpen
+	}
+	runtime.KeepAlive(g)
 	return ret, nil
 }