The less C code we have, the better. This PR moves the GCM Seal/Open implementation from C to Go. It was previously written in C to avoid allocations, but that's no longer need thanks to the `noescape` and `nocallback` directive.

These benchmark run demonstrates that there is not performance regression:

```
goos: linux
goarch: amd64
pkg: github.com/golang-fips/openssl/v2
cpu: Intel(R) Core(TM) i7-10850H CPU @ 2.70GHz
              │   old.txt    │               new.txt               │
              │    sec/op    │    sec/op     vs base               │
AESGCM_Open-4   1.708µ ± 17%   1.814µ ±  6%       ~ (p=0.190 n=10)
AESGCM_Seal-4   1.876µ ± 21%   1.771µ ± 15%       ~ (p=0.224 n=10)
geomean         1.790µ         1.792µ        +0.13%

              │    old.txt    │               new.txt                │
              │      B/s      │      B/s       vs base               │
AESGCM_Open-4   35.74Mi ± 14%   33.65Mi ±  5%       ~ (p=0.190 n=10)
AESGCM_Seal-4   32.57Mi ± 17%   34.47Mi ± 13%       ~ (p=0.224 n=10)
geomean         34.12Mi         34.06Mi        -0.18%

              │   old.txt    │               new.txt               │
              │     B/op     │    B/op     vs base                 │
AESGCM_Open-4   0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
AESGCM_Seal-4   0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
geomean                    ²               +0.00%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean

              │   old.txt    │               new.txt               │
              │  allocs/op   │ allocs/op   vs base                 │
AESGCM_Open-4   0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
AESGCM_Seal-4   0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
geomean                    ²               +0.00%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean
```